Add tests for resume resource
