+++ backend/tests/resources/base.py
@@ -1,0 +1,10 @@
+import pytest
+
+from app import app
+
+
+@pytest.fixture
+def client():
+    app.config["TESTING"] = True
+    with app.test_client() as client:
+        yield client
