+++ backend/tests/resources/test_hello.py
@@ -1,13 +1,4 @@
-import pytest
-
-from app import app
-
-
-@pytest.fixture
-def client():
-    app.config["TESTING"] = True
-    with app.test_client() as client:
-        yield client
+from .base import client
 
 
 def test_get(client):
